Split CSV entry into two modules
